+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Added `options.WithAddAttribute` and `options.WithDropAttribute` options for `session.AlterTable` request
+* Added return error while create topic writer with not equal producer id and message group id.
 
 ## v3.39.0
 * Removed message level partitioning from experimental topic API. It is unavailable on server side yet.